Created index.html and style.css
Using the project and the landing page provided in Odin Project, I began using
flex display to organize the elements in the index to format the page to what
the project is looking for. Then I began messing around with the style page to
work with flex display and to also just overall begin getting a better feel for
how flex works in css and html.